Replace submodules with downloaded scripts
There's been some issues with running the submodules, as they're designed
to be run in a built form, so now the build system grabs a built copy from
the appropriate external locations. Specifically, PHP Code Sniffer requires
custom code standards to be put in a location inside itself, which
submodules can not handle.